moved compared_with_golden functions into file /forge/verify/compare.py -> solved circular dependency issues
