+++ forge/forge/op/eval/__init__.py
@@ -4,11 +4,9 @@
 from .common import (
     dump_tensor,
     eval_debug_print,
-    compare_tensor_to_golden,
     create_constant_tensor_from_tile,
     create_constant_tensor_from_value,
     create_constant_tensor_from_tensor,
-    calculate_pcc,
 )
 from .sparse_utils import (
     create_flattened_padding_removal_sparse_picker_matrix,
