+++ forge/forge/python_codegen.py
@@ -154,6 +154,7 @@ class ForgeWriter(PythonWriter):
         if include_pytest_imports:
             self.wl("")
             self.wl("from forge import Tensor, compile")
+            self.wl("from forge.verify.compare import compare_with_golden")
             self.wl("from forge.verify.verify import verify")
             self.wl("from forge.verify.config import VerifyConfig")
             self.wl("import pytest")
