+++ forge/test/benchmark/benchmark/models/mnist_linear.py
@@ -10,7 +10,7 @@ import json
 import torch
 from torch import nn
 import forge
-from forge.op.eval.common import compare_with_golden_pcc
+from forge.verify.compare import compare_with_golden_pcc
 
 
 # Batch size configurations
