+++ forge/test/mlir/test_loss.py
@@ -7,7 +7,7 @@ import torch
 from torch import nn
 
 import forge
-from forge.op.eval.common import compare_with_golden
+from forge.verify.compare import compare_with_golden
 
 
 @pytest.mark.parametrize(
