+++ forge/test/mlir/test_training.py
@@ -8,7 +8,7 @@ import pytest
 
 import forge
 import forge.config
-from forge.op.eval.common import compare_with_golden
+from forge.verify.compare import compare_with_golden
 
 
 @pytest.mark.push
