+++ forge/test/models/pytorch/multimodal/vilt/test_vilt.py
@@ -10,7 +10,7 @@ import torch
 from PIL import Image
 from transformers import ViltProcessor, ViltForQuestionAnswering, ViltForMaskedLM, ViltConfig
 from test.models.pytorch.multimodal.vilt.utils.model import ViLtEmbeddingWrapper, ViltModelWrapper
-from forge.op.eval.common import compare_with_golden
+from forge.verify.compare import compare_with_golden
 
 url = "http://images.cocodataset.org/val2017/000000039769.jpg"
 image = Image.open(requests.get(url, stream=True).raw)
