+++ forge/test/models/pytorch/text/codegen/test_codegen.py
@@ -16,7 +16,7 @@ variants = [
     # "Salesforce/codegen-350M-nl", # Currently not supported
 ]
 import torch
-from forge.op.eval.common import compare_with_golden
+from forge.verify.compare import compare_with_golden
 
 
 @pytest.mark.nightly
