+++ forge/test/models/pytorch/text/t5/test_t5.py
@@ -10,7 +10,7 @@ import forge
 import torch
 from forge.transformers.pipeline import pipeline as forge_pipeline
 from transformers import T5ForConditionalGeneration, T5Tokenizer, T5Config
-from forge.op.eval.common import compare_with_golden
+from forge.verify.compare import compare_with_golden
 
 
 @pytest.mark.nightly
